Refactor: Move MultiplayerGameViewModel and related files to subpackage
Moves `MultiplayerGameViewModel` and its associated screen components (`LobbyScreen.kt`, `PlayerList.kt`) into a new `multiplayer` subpackage within `ui.main`.

Also moves `MultiplayerGameManager` from the `models` package to the `services` package to better reflect its role.

Updates import statements accordingly.
