+++ app/src/main/java/com/example/infiltrados/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.example.infiltrados
 
-import android.content.Context
 import android.util.Log
 
 import android.os.Bundle
@@ -16,24 +15,20 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import androidx.navigation.compose.rememberNavController
-import androidx.navigation.toRoute
 import com.example.infiltrados.services.GameManager
 import com.example.infiltrados.services.WordLoader
 import com.example.infiltrados.ui.main.DiscussionScreen
 import com.example.infiltrados.ui.main.EndGameScreen
 import com.example.infiltrados.ui.main.LobbyScreen
 import com.example.infiltrados.ui.main.MrWhiteGuessScreen
-import com.example.infiltrados.ui.main.MultiplayerGameViewModel
+import com.example.infiltrados.ui.main.multiplayer.MultiplayerGameViewModel
 import com.example.infiltrados.ui.main.PlayerEliminatedScreen
 import com.example.infiltrados.ui.main.PlayerInputScreen
 import com.example.infiltrados.ui.main.VotationScreen
 import com.example.infiltrados.ui.main.WordRevealScreen
 import com.example.infiltrados.ui.main.RulesScreen
 import com.example.infiltrados.ui.main.SplashScreen
-import com.example.infiltrados.ui.main.multiplayerLobbyScreen.OnlineLobbyScreen
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
+import com.example.infiltrados.ui.main.multiplayer.multiplayerLobbyScreen.OnlineLobbyScreen
 import kotlinx.serialization.Serializable
